Renamed luminosity to luminance.
Renamed LuminosityMaterial to LuminanceMaterial and AdaptiveLuminosityMaterial to AdaptiveLuminanceMaterial.
