+++ src/materials/glsl/adaptive-luminance/shader.vert
