+++ src/materials/glsl/luminance/shader.frag
