+++ src/materials/index.js
@@ -4,7 +4,7 @@
  * @module postprocessing/materials
  */
 
-export { AdaptiveLuminosityMaterial } from "./AdaptiveLuminosityMaterial.js";
+export { AdaptiveLuminanceMaterial } from "./AdaptiveLuminanceMaterial.js";
 export { BokehMaterial } from "./BokehMaterial.js";
 export { ColorEdgesMaterial } from "./ColorEdgesMaterial.js";
 export { CombineMaterial } from "./CombineMaterial.js";
@@ -16,7 +16,7 @@ export { EffectMaterial } from "./EffectMaterial.js";
 export { FilmMaterial } from "./FilmMaterial.js";
 export { GlitchMaterial } from "./GlitchMaterial.js";
 export { GodRaysMaterial } from "./GodRaysMaterial.js";
-export { LuminosityMaterial } from "./LuminosityMaterial.js";
+export { LuminanceMaterial } from "./LuminanceMaterial.js";
 export { OutlineBlendMaterial } from "./OutlineBlendMaterial.js";
 export { OutlineEdgesMaterial } from "./OutlineEdgesMaterial.js";
 export { PixelationMaterial } from "./PixelationMaterial.js";
